backend/requirements.txt
flask
python-dotenv
openai
PyPDF2
